🛠️ refactor(comment): controller, router 폴더 이름 획일화
